+++ src/easysam/inspect.py
@@ -11,8 +11,19 @@ from easysam.validate_cloud import validate as validate_cloud
 
 
 @click.group(help='Inspect the application for debugging purposes')
-def inspect():
-    pass
+@click.pass_obj
+@click.option('--environment', type=str)
+@click.option('--target-region', type=str)
+def inspect(obj, environment, target_region):
+    deploy_ctx = {}
+
+    if environment:
+        deploy_ctx['environment'] = environment
+
+    if target_region:
+        deploy_ctx['region'] = target_region
+
+    obj.update({'deploy_ctx': deploy_ctx})
 
 
 @inspect.command(name='common-deps', help='Inspect a lambda function')
@@ -35,13 +46,15 @@ def common_deps(common_dir, lambda_dir):
 
 
 @inspect.command(help='Validate the resources.yaml file')
+@click.pass_obj
 @click.option('--path', multiple=True)
 @click.argument('directory', type=click.Path(exists=True))
-def schema(directory, path):
+def schema(obj, directory, path):
     directory = Path(directory)
     pypath = [Path(p) for p in path]
     errors = []
-    resources_data = load_resources(directory, pypath, {}, errors)
+    deploy_ctx = obj['deploy_ctx']
+    resources_data = load_resources(directory, pypath, deploy_ctx, errors)
 
     if errors:
         rich.print(f'[red]There were {len(errors)} validation errors.[/red]')
@@ -62,7 +75,12 @@ def cloud(obj, directory, path, environment):
     directory = Path(directory)
     pypath = [Path(p) for p in path]
     errors = []
-    resources_data = load_resources(directory, pypath, {}, errors)
+    deploy_ctx = obj['deploy_ctx']
+
+    if 'environment' not in obj:
+        raise click.UsageError('Environment is required for cloud inspection')
+
+    resources_data = load_resources(directory, pypath, deploy_ctx, errors)
 
     if errors:
         rich.print(