+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.3.0
 
+- Added conditional resources support
 - Added `inspect schema` command to validate the resources.yaml file
 - Added `inspect cloud` command to inspect required cloud resources